+++ pkg/unarchiver/unarchiver_test.go
@@ -7,7 +7,7 @@ import (
 	"sort"
 	"testing"
 
-	"github.com/ActiveState/cli/internal/unarchiver"
+	"github.com/ActiveState/cli/pkg/unarchiver"
 	"github.com/stretchr/testify/suite"
 )
